control-plane/pyproject.toml
[project]
name = "control-plane"
version = "0.1.0"
description = "The Brain of Zero Trust"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "sqlalchemy>=2.0.25",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    "requests>=2.32.5",
    "qrcode[pil]>=7.4.2",
    "pillow>=10.2.0",
]
